# internal/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "whois",
    "description": "WHOIS/RDAP query service. Returns registration data for domain names, IP networks (including CIDR prefixes) and autonomous system numbers as JSON aligned with the RDAP vocabulary (RFC 9083). Errors follow RFC 9457 Problem Details. API key authentication is disabled by default; when the operator configures `auth.keys`, every endpoint except /health and /ready requires a key sent as `Authorization: Bearer <key>` or `X-API-Key: <key>`.",
    "version": "0.9.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/domain/{name}": {
      "get": {
        "operationId": "queryDomain",
        "summary": "Query a domain name (RFC 9082-style typed path)",
        "description": "Returns 400 when the resource is not a valid domain name instead of auto-detecting another type.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "Domain name; Unicode (IDN) names are accepted.",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/raw"
          },
          {
            "$ref": "#/components/parameters/refresh"
          },
          {
            "$ref": "#/components/parameters/ifNoneMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "Domain registration data, or the raw WHOIS text when `raw` is set.",
            "headers": {
              "X-Cache": {
                "$ref": "#/components/headers/X-Cache"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/Cache-Control"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Domain"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Unparsed WHOIS response (only with `?raw=1`)."
                }
              }
            }
          },
          "304": {
            "$ref": "#/components/responses/NotModified"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/QueryDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/ip/{address}": {
      "get": {
        "operationId": "queryIP",
        "summary": "Query an IP network (RFC 9082-style typed path)",
        "description": "Accepts a single IPv4/IPv6 address or a CIDR prefix whose slash is part of the path, e.g. `/ip/192.0.2.0/24` or `/ip/2001:db8::/32`. Returns 400 when the resource is not a valid IP or prefix.",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "IP address (`192.0.2.1`) or CIDR prefix (`192.0.2.0/24`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/refresh"
          },
          {
            "$ref": "#/components/parameters/ifNoneMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "IP network registration data.",
            "headers": {
              "X-Cache": {
                "$ref": "#/components/headers/X-Cache"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/Cache-Control"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IPNetwork"
                }
              }
            }
          },
          "304": {
            "$ref": "#/components/responses/NotModified"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/QueryDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/autnum/{asn}": {
      "get": {
        "operationId": "queryAutnum",
        "summary": "Query an autonomous system number (RFC 9082-style typed path)",
        "description": "Returns 400 when the resource is not a valid AS number.",
        "parameters": [
          {
            "name": "asn",
            "in": "path",
            "required": true,
            "description": "AS number, with optional case-insensitive `as`/`asn` prefix (`205794`, `as205794`, `AS205794`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/refresh"
          },
          {
            "$ref": "#/components/parameters/ifNoneMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "Autnum registration data.",
            "headers": {
              "X-Cache": {
                "$ref": "#/components/headers/X-Cache"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/Cache-Control"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Autnum"
                }
              }
            }
          },
          "304": {
            "$ref": "#/components/responses/NotModified"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/QueryDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/batch": {
      "post": {
        "operationId": "batchQuery",
        "summary": "Query multiple domains, IPs/CIDR prefixes, or ASNs in one request",
        "description": "Answers each query independently and always returns 200 with per-item statuses: successful items carry the regular response object in `data`, failed items an RFC 9457 problem object in `error`. Disabled by default; the operator enables it with `batch.enabled`, and the number of queries per request is capped by `batch.maxItems` (default 10). When the API key has a per-key rate limit, a batch of N queries costs N tokens — the same as N single queries.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "queries"
                ],
                "properties": {
                  "queries": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Domain names (IDN accepted), IP addresses, CIDR prefixes, or ASNs, freely mixed."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Per-item results in request order.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Batch queries are disabled on this instance (problem type `batch-disabled`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/info": {
      "get": {
        "operationId": "info",
        "summary": "Runtime information (version, uptime, Go version)",
        "responses": {
          "200": {
            "description": "Runtime information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/openapi.json": {
      "get": {
        "operationId": "openapi",
        "summary": "This OpenAPI document",
        "parameters": [
          {
            "$ref": "#/components/parameters/ifNoneMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "OpenAPI 3.1 description of this service.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "304": {
            "$ref": "#/components/responses/NotModified"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/mcp": {
      "post": {
        "operationId": "mcp",
        "summary": "MCP Streamable HTTP endpoint",
        "description": "Model Context Protocol endpoint exposing the `whois_lookup` tool for AI assistants. The message format is defined by the MCP specification, not this document.",
        "externalDocs": {
          "description": "Model Context Protocol",
          "url": "https://modelcontextprotocol.io"
        },
        "responses": {
          "200": {
            "description": "MCP protocol response."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key sent as `Authorization: Bearer <key>`. Only enforced when the server is configured with `auth.keys`; authentication is disabled by default."
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key sent as `X-API-Key: <key>`. Only enforced when the server is configured with `auth.keys`; authentication is disabled by default."
      }
    },
    "parameters": {
      "ifNoneMatch": {
        "name": "If-None-Match",
        "in": "header",
        "required": false,
        "description": "Conditional revalidation (RFC 9110): when the value matches the current `ETag` of the response, the server answers `304 Not Modified` with no body.",
        "schema": {
          "type": "string"
        }
      },
      "raw": {
        "name": "raw",
        "in": "query",
        "required": false,
        "description": "Return the unparsed WHOIS response as text/plain (domains only; IP and ASN lookups use RDAP, which has no raw-text form). `raw=0` and `raw=false` opt out; any other presence of the parameter opts in.",
        "schema": {
          "type": "string"
        }
      },
      "refresh": {
        "name": "refresh",
        "in": "query",
        "required": false,
        "description": "Bypass the server-side cache and query the upstream registry, overwriting the cached entry (response carries `X-Cache: REFRESH`). Only honored when API key authentication is enabled; open instances return 403 `refresh-requires-auth`. `refresh=0` and `refresh=false` opt out; any other presence of the parameter opts in.",
        "schema": {
          "type": "string"
        }
      }
    },
    "headers": {
      "ETag": {
        "description": "Strong entity tag of the response body, usable for `If-None-Match` revalidation.",
        "schema": {
          "type": "string"
        }
      },
      "X-Cache": {
        "description": "Whether the response was served from the server-side cache (HIT), fetched from the upstream registry (MISS), or forced upstream by `?refresh` (REFRESH).",
        "schema": {
          "type": "string",
          "enum": [
            "HIT",
            "MISS",
            "REFRESH"
          ]
        }
      },
      "Cache-Control": {
        "description": "Successful responses are cacheable for the server's configured cache expiration: `public, max-age=<seconds>`.",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
...
      "NotModified": {
        "description": "The entity tag in `If-None-Match` matches the current response; the body is unchanged and omitted.",
        "headers": {
          "ETag": {
            "$ref": "#/components/headers/ETag"
          }
        }
      },
      "Unauthorized": {
        "description": "API key authentication is enabled on this server and the request carried no valid key. Only returned when the operator configures `auth.keys`; authentication is disabled by default.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "BadRequest": {
        "description": "The input is not a valid domain, IP, CIDR prefix, or ASN — or the parameter combination is unsupported.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource is not registered or the registry returned no data.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "QueryDenied": {
        "description": "The upstream registry refused to answer the query (problem type `query-denied`), or `?refresh` was used on an instance without API key authentication enabled (problem type `refresh-requires-auth`).",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "RateLimited": {
        "description": "The server's concurrent-request limit was reached (retry after a short delay), or the API key's per-key rate limit is exhausted (per-key rejections carry a Retry-After header).",
        "headers": {
          "Retry-After": {
            "description": "Seconds until the next request is allowed. Only present on per-key rate limit rejections.",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "ServerError": {
        "description": "The upstream query failed or an unexpected server-side error occurred.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
      "BatchResponse": {
        "type": "object",
        "required": [
          "results"
        ],
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchItem"
            },
            "description": "One entry per query, in request order."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 Problem Details. `type` and `title` are stable identifiers; `detail` is human-readable and may change between releases. See https://github.com/KincaidYang/whois/blob/main/docs/errors.md",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "URI of the problem type's documentation anchor."
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "Domain": {
        "type": "object",
        "description": "Domain registration data; field names follow the RDAP domain object (RFC 9083 section 5.3). Dates are RFC 3339 UTC, or `YYYY-MM-DD` when the registry publishes only a date.",
        "required": [
          "objectClassName",
          "status",
          "nameservers"
        ],
        "properties": {
          "objectClassName": {
            "type": "string",
            "const": "domain"
          },
          "ldhName": {
            "type": "string",
            "description": "Lowercase ASCII (punycode) domain name."
          },
          "unicodeName": {
            "type": "string",
            "description": "Unicode form of the name; present for IDN domains."
          },
          "registrar": {
            "type": "string"
          },
          "registrarIanaId": {
            "type": "string"
          },
          "status": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Status values with ICANN EPP reference URLs stripped and duplicates removed."
          },
          "registrationDate": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string"
          },
          "lastChangedDate": {
            "type": "string"
          },
          "nameservers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lowercase nameserver host names."
          },
          "secureDNS": {
            "$ref": "#/components/schemas/SecureDNS"
          },
          "lastUpdateOfRdapDb": {
            "type": "string"
          },
          "unparsed": {
            "type": "boolean",
            "description": "True when no parser exists for the TLD; the response carries the raw WHOIS text in rawText instead of parsed fields."
          },
          "rawText": {
            "type": "string",
            "description": "Raw WHOIS response (only when unparsed is true)."
          }
        }
      },
...
      "IPNetwork": {
        "type": "object",
        "description": "IP network registration data; field names follow the RDAP ip network object (RFC 9083 section 5.4).",
        "required": [
          "objectClassName",
          "handle",
          "status"
        ],
        "properties": {
          "objectClassName": {
            "type": "string",
            "const": "ip network"
          },
          "handle": {
            "type": "string"
          },
          "startAddress": {
            "type": "string"
          },
          "endAddress": {
            "type": "string"
          },
          "cidr": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "status": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "registrationDate": {
            "type": "string"
          },
          "lastChangedDate": {
            "type": "string"
          },
          "remarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Remark"
            }
          }
        }
      },
      "Autnum": {
        "type": "object",
        "description": "Autonomous system registration data; field names follow the RDAP autnum object (RFC 9083 section 5.5).",
        "required": [
          "objectClassName",
          "handle",
          "status"
        ],
        "properties": {
          "objectClassName": {
            "type": "string",
            "const": "autnum"
          },
          "handle": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "registrationDate": {
            "type": "string"
          },
          "lastChangedDate": {
            "type": "string"
          },
          "remarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Remark"
            }
          }
        }
      },
...
    }
  }
}